+++ src/components/Lyric.js
@@ -5,7 +5,7 @@ import TextField from '@mui/material/TextField';
 import { withStyles } from '@mui/styles';
 import Grid from '@mui/material/Grid';
 import { ScrollBar } from '../styles/styles';
-import LyricSearchBar from './LyricSearchBar';
+import LyricSearchBar from './lyric/LyricSearchBar';
 import { StorageManagerCtx } from '../contexts/StorageManagerContext';
 import { skinPreset } from '../styles/skin';
