+++ src/components/lyric/LyricSearchBar.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect, useContext } from 'react';
 import TextField from '@mui/material/TextField';
 import Autocomplete from '@mui/material/Autocomplete';
-import { searchLyricOptions, searchLyric } from '../utils/Data';
-import { StorageManagerCtx } from '../contexts/StorageManagerContext';
+import { searchLyricOptions, searchLyric } from '../../utils/Data';
+import { StorageManagerCtx } from '../../contexts/StorageManagerContext';
 
 export default function LyricSearchBar({
   SearchKey,
